Created Custom Video Parser
This file consumes the data specified in the wlasl_dataset folder and outputs frames of the video in the processed_frames_dataset/ folder.
Each folder is a label, and has numerous sub-folders with a unique id. All images corresponding to that video are present in the folder.
The annotations.txt file is a line by line meta data that denotes "VideoPath/Location StartFrame EndFrame Label_ID"